sync desktop menu and mobile main navigation
